+++ survey_cad_truck_gui/src/io_utils.rs
@@ -1,0 +1,54 @@
+use survey_cad::geometry::{Arc, Point};
+
+pub fn read_line_csv(path: &str, dst_epsg: u32) -> std::io::Result<(Point, Point)> {
+    let pts = survey_cad::io::read_points_csv(path, Some(4326), Some(dst_epsg))?;
+    if pts.len() != 2 {
+        return Err(std::io::Error::new(
+            std::io::ErrorKind::InvalidData,
+            "expected exactly two points",
+        ));
+    }
+    Ok((pts[0], pts[1]))
+}
+
+pub fn read_points_list(path: &str, dst_epsg: u32) -> std::io::Result<Vec<Point>> {
+    survey_cad::io::read_points_csv(path, Some(4326), Some(dst_epsg))
+}
+
+pub fn read_arc_csv(path: &str) -> std::io::Result<Arc> {
+    let lines = survey_cad::io::read_lines(path)?;
+    if lines.is_empty() {
+        return Err(std::io::Error::new(
+            std::io::ErrorKind::InvalidData,
+            "empty file",
+        ));
+    }
+    let parts: Vec<&str> = lines[0].split(',').collect();
+    if parts.len() != 5 {
+        return Err(std::io::Error::new(
+            std::io::ErrorKind::InvalidData,
+            "expected cx,cy,radius,start,end",
+        ));
+    }
+    let cx: f64 = parts[0]
+        .trim()
+        .parse()
+        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
+    let cy: f64 = parts[1]
+        .trim()
+        .parse()
+        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
+    let r: f64 = parts[2]
+        .trim()
+        .parse()
+        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
+    let sa: f64 = parts[3]
+        .trim()
+        .parse()
+        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
+    let ea: f64 = parts[4]
+        .trim()
+        .parse()
+        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
+    Ok(Arc::new(Point::new(cx, cy), r, sa, ea))
+}
