+++ survey_cad_truck_gui/src/main.rs
@@ -43,8 +43,16 @@ mod ui_state;
 mod commands;
 mod python;
 mod render;
+mod workspace;
+mod cross_section;
+mod io_utils;
+mod inspector;
 
 use once_cell::sync::Lazy;
+pub use workspace::{render_workspace, refresh_workspace, set_workspace_image_result};
+pub use cross_section::{render_cross_section, calc_section_params, screen_to_world, nearest_point, grade_at, SectionParams};
+pub use io_utils::{read_line_csv, read_points_list, read_arc_csv};
+pub use inspector::{show_context_menu, has_selection, show_inspector_for_point, show_inspector_for_polygon};
 use commands::{Command, CommandStack, Context, MacroPlaying, MacroRecorder, record_macro, spawn_line, spawn_point};
 use python::{MacroContext, PythonContext, play_macro_file, run_python_file};
 use render::{
@@ -69,1583 +77,6 @@ static FONT_DATA: &[u8] = include_bytes!(env!("DEFAULT_FONT_PATH"));
 static FONT: Lazy<Font<'static>> =
     Lazy::new(|| Font::try_from_bytes(FONT_DATA).unwrap());
 
-
-fn render_workspace(
-    data: &WorkspaceRenderData,
-    state: &RenderState,
-    styles: &RenderStyles,
-    drawing: &DrawingMode,
-    grid: &GridSettings,
-    width: u32,
-    height: u32,
-) -> Result<Image, Box<dyn Error>> {
-    if width == 0 || height == 0 {
-        // When the window has not been laid out yet slint reports a size of
-        // zero. Returning an empty image avoids panicking until a real size is
-        // available.
-        return Ok(Image::default());
-    }
-    let mut pixmap = state.pixmap.borrow_mut();
-    if pixmap.width() != width || pixmap.height() != height {
-        *pixmap = Pixmap::new(width, height).ok_or("failed to create pixmap")?;
-    } else {
-        pixmap.fill(Color::from_rgba8(32, 32, 32, 255));
-    }
-    let mut paint = Paint::default();
-    paint.set_color(Color::from_rgba8(
-        grid.color[0],
-        grid.color[1],
-        grid.color[2],
-        255,
-    ));
-    paint.anti_alias = true;
-    let grid_stroke = Stroke {
-        width: 1.0,
-        ..Stroke::default()
-    };
-    let origin_x = width as f32 / 2.0;
-    let origin_y = height as f32 / 2.0;
-    let zoom_val = *state.zoom.borrow();
-    let off = state.offset.borrow();
-    let off_x = off.x;
-    let off_y = off.y;
-    drop(off);
-    let tx = |x: f32| (x + off_x) * zoom_val + origin_x;
-    let ty = |y: f32| origin_y - (y + off_y) * zoom_val;
-    let step = grid.spacing * zoom_val;
-    let mut x = origin_x;
-    if grid.visible {
-        while x < width as f32 {
-            let mut pb = PathBuilder::new();
-            pb.move_to(x, 0.0);
-            pb.line_to(x, height as f32);
-            if let Some(p) = pb.finish() {
-                pixmap.stroke_path(&p, &paint, &grid_stroke, Transform::identity(), None);
-            }
-            x += step;
-        }
-        x = origin_x - step;
-        while x >= 0.0 {
-            let mut pb = PathBuilder::new();
-            pb.move_to(x, 0.0);
-            pb.line_to(x, height as f32);
-            if let Some(p) = pb.finish() {
-                pixmap.stroke_path(&p, &paint, &grid_stroke, Transform::identity(), None);
-            }
-            x -= step;
-        }
-        let mut y = origin_y;
-        while y < height as f32 {
-            let mut pb = PathBuilder::new();
-            pb.move_to(0.0, y);
-            pb.line_to(width as f32, y);
-            if let Some(p) = pb.finish() {
-                pixmap.stroke_path(&p, &paint, &grid_stroke, Transform::identity(), None);
-            }
-            y += step;
-        }
-        y = origin_y - step;
-        while y >= 0.0 {
-            let mut pb = PathBuilder::new();
-            pb.move_to(0.0, y);
-            pb.line_to(width as f32, y);
-            if let Some(p) = pb.finish() {
-                pixmap.stroke_path(&p, &paint, &grid_stroke, Transform::identity(), None);
-            }
-            y -= step;
-        }
-    }
-    paint.set_color(Color::from_rgba8(90, 90, 90, 255));
-    let mut pb = PathBuilder::new();
-    pb.move_to(origin_x, 0.0);
-    pb.line_to(origin_x, height as f32);
-    if let Some(path) = pb.finish() {
-        pixmap.stroke_path(&path, &paint, &grid_stroke, Transform::identity(), None);
-    }
-    let mut pb = PathBuilder::new();
-    pb.move_to(0.0, origin_y);
-    pb.line_to(width as f32, origin_y);
-    if let Some(path) = pb.finish() {
-        pixmap.stroke_path(&path, &paint, &grid_stroke, Transform::identity(), None);
-    }
-
-    paint.set_color(Color::from_rgba8(255, 0, 0, 255));
-    for (i, (s, e)) in data.lines.iter().enumerate() {
-        let selected = state
-            .selected_lines
-            .borrow()
-            .iter()
-            .any(|(ls, le)| (*ls == *s && *le == *e) || (*ls == *e && *le == *s));
-        let style_idx = styles
-            .line_style_indices
-            .borrow()
-            .get(i)
-            .copied()
-            .unwrap_or(0);
-        let mut style = styles
-            .line_styles
-            .get(style_idx)
-            .copied()
-            .unwrap_or_default();
-        if selected {
-            style.color = [255, 255, 0];
-        }
-        paint.set_color(Color::from_rgba8(
-            style.color[0],
-            style.color[1],
-            style.color[2],
-            255,
-        ));
-        let mut stroke = Stroke {
-            width: style.weight.0,
-            ..Stroke::default()
-        };
-        use tiny_skia::StrokeDash;
-        match style.line_type {
-            LineType::Dashed => {
-                stroke.dash = StrokeDash::new(vec![10.0, 10.0], 0.0);
-            }
-            LineType::Dotted => {
-                stroke.dash = StrokeDash::new(vec![2.0, 6.0], 0.0);
-            }
-            _ => {}
-        }
-        let mut pb = PathBuilder::new();
-        pb.move_to(tx(s.x as f32), ty(s.y as f32));
-        pb.line_to(tx(e.x as f32), ty(e.y as f32));
-        if let Some(path) = pb.finish() {
-            let stroke = Stroke {
-                width: 1.0,
-                ..Stroke::default()
-            };
-            pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-        }
-
-        if styles.show_labels {
-            let line = Line::new(*s, *e);
-            let ann = LineAnnotation::from_line(&line);
-            let mut angle = 90.0 - ann.azimuth.to_degrees();
-            if angle < 0.0 {
-                angle += 360.0;
-            }
-            let text = format!("{:.2} m\n{}", ann.distance, format_dms(angle));
-            let mid = line.midpoint();
-            let dx = e.x - s.x;
-            let dy = e.y - s.y;
-            let len = (dx * dx + dy * dy).sqrt();
-            let (ox, oy) = if len > 0.0 {
-                let nx = dx / len;
-                let ny = dy / len;
-                match styles.label_style.position {
-                    LineLabelPosition::Above => (-ny as f32, nx as f32),
-                    LineLabelPosition::Below => (ny as f32, -nx as f32),
-                    LineLabelPosition::Center => (0.0, 0.0),
-                }
-            } else {
-                (0.0, 0.0)
-            };
-            draw_text(
-                &mut pixmap,
-                &text,
-                &FONT,
-                tx(mid.x as f32 + ox * 10.0),
-                ty(mid.y as f32 + oy * 10.0),
-                Color::from_rgba8(
-                    styles.label_style.color[0],
-                    styles.label_style.color[1],
-                    styles.label_style.color[2],
-                    255,
-                ),
-                styles.label_style.text_style.height as f32,
-            );
-        }
-    }
-
-    for (i, poly) in data.polygons.iter().enumerate() {
-        if poly.len() < 2 {
-            continue;
-        }
-        let mut pb = PathBuilder::new();
-        let Some(first) = poly.first() else { continue };
-        pb.move_to(tx(first.x as f32), ty(first.y as f32));
-        for p in &poly[1..] {
-            pb.line_to(tx(p.x as f32), ty(p.y as f32));
-        }
-        pb.close();
-        if let Some(path) = pb.finish() {
-            let style_idx = styles
-                .polygon_style_indices
-                .borrow()
-                .get(i)
-                .copied()
-                .unwrap_or(0);
-            let pstyle = styles
-                .polygon_styles
-                .get(style_idx)
-                .copied()
-                .unwrap_or_default();
-
-            paint.set_color(Color::from_rgba8(
-                pstyle.fill_color[0],
-                pstyle.fill_color[1],
-                pstyle.fill_color[2],
-                255,
-            ));
-            pixmap.fill_path(
-                &path,
-                &paint,
-                FillRule::Winding,
-                Transform::identity(),
-                None,
-            );
-
-            if pstyle.hatch_pattern != HatchPattern::None {
-                paint.set_color(Color::from_rgba8(
-                    pstyle.hatch_color[0],
-                    pstyle.hatch_color[1],
-                    pstyle.hatch_color[2],
-                    255,
-                ));
-                let stroke = Stroke {
-                    width: 1.0,
-                    ..Stroke::default()
-                };
-                {
-                    let bb = path.bounds();
-                    let step = 10.0;
-                    if matches!(
-                        pstyle.hatch_pattern,
-                        HatchPattern::Cross | HatchPattern::Grid
-                    ) {
-                        let mut x = bb.left();
-                        while x <= bb.right() {
-                            let mut pb = PathBuilder::new();
-                            pb.move_to(x, bb.top());
-                            pb.line_to(x, bb.bottom());
-                            if let Some(p) = pb.finish() {
-                                pixmap.stroke_path(
-                                    &p,
-                                    &paint,
-                                    &stroke,
-                                    Transform::identity(),
-                                    None,
-                                );
-                            }
-                            x += step;
-                        }
-                        let mut y = bb.top();
-                        while y <= bb.bottom() {
-                            let mut pb = PathBuilder::new();
-                            pb.move_to(bb.left(), y);
-                            pb.line_to(bb.right(), y);
-                            if let Some(p) = pb.finish() {
-                                pixmap.stroke_path(
-                                    &p,
-                                    &paint,
-                                    &stroke,
-                                    Transform::identity(),
-                                    None,
-                                );
-                            }
-                            y += step;
-                        }
-                    }
-                    if pstyle.hatch_pattern == HatchPattern::ForwardDiagonal {
-                        let mut x = bb.left() - bb.height();
-                        while x <= bb.right() {
-                            let mut pb = PathBuilder::new();
-                            pb.move_to(x, bb.bottom());
-                            pb.line_to(x + bb.height(), bb.top());
-                            if let Some(p) = pb.finish() {
-                                pixmap.stroke_path(
-                                    &p,
-                                    &paint,
-                                    &stroke,
-                                    Transform::identity(),
-                                    None,
-                                );
-                            }
-                            x += step;
-                        }
-                    } else if pstyle.hatch_pattern == HatchPattern::BackwardDiagonal {
-                        let mut x = bb.left();
-                        while x <= bb.right() + bb.height() {
-                            let mut pb = PathBuilder::new();
-                            pb.move_to(x, bb.top());
-                            pb.line_to(x - bb.height(), bb.bottom());
-                            if let Some(p) = pb.finish() {
-                                pixmap.stroke_path(
-                                    &p,
-                                    &paint,
-                                    &stroke,
-                                    Transform::identity(),
-                                    None,
-                                );
-                            }
-                            x += step;
-                        }
-                    }
-                }
-            }
-
-            let selected = state.selected_polygons.borrow().contains(&i);
-            if selected {
-                paint.set_color(Color::from_rgba8(255, 255, 0, 255));
-            } else {
-                paint.set_color(Color::from_rgba8(255, 0, 0, 255));
-            }
-            let stroke = Stroke {
-                width: 1.0,
-                ..Stroke::default()
-            };
-            pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-        }
-    }
-
-    for (i, pl) in data.polylines.iter().enumerate() {
-        if pl.vertices.len() < 2 {
-            continue;
-        }
-        let mut pb = PathBuilder::new();
-        let first = &pl.vertices[0];
-        pb.move_to(tx(first.x as f32), ty(first.y as f32));
-        for p in &pl.vertices[1..] {
-            pb.line_to(tx(p.x as f32), ty(p.y as f32));
-        }
-        if let Some(path) = pb.finish() {
-            let selected = state.selected_polylines.borrow().contains(&i);
-            if selected {
-                paint.set_color(Color::from_rgba8(255, 255, 0, 255));
-            }
-            let stroke = Stroke {
-                width: 1.0,
-                ..Stroke::default()
-            };
-            pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-            if selected {
-                paint.set_color(Color::from_rgba8(255, 0, 0, 255));
-            }
-        }
-    }
-
-    for (i, arc) in data.arcs.iter().enumerate() {
-        let steps = 32;
-        let mut pb = PathBuilder::new();
-        for i in 0..=steps {
-            let t = arc.start_angle + (arc.end_angle - arc.start_angle) * (i as f64 / steps as f64);
-            let x = arc.center.x + arc.radius * t.cos();
-            let y = arc.center.y + arc.radius * t.sin();
-            let px = tx(x as f32);
-            let py = ty(y as f32);
-            if i == 0 {
-                pb.move_to(px, py);
-            } else {
-                pb.line_to(px, py);
-            }
-        }
-        if let Some(path) = pb.finish() {
-            let selected = state.selected_arcs.borrow().contains(&i);
-            if selected {
-                paint.set_color(Color::from_rgba8(255, 255, 0, 255));
-            }
-            let stroke = Stroke {
-                width: 1.0,
-                ..Stroke::default()
-            };
-            pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-            if selected {
-                paint.set_color(Color::from_rgba8(255, 0, 0, 255));
-            }
-        }
-    }
-
-    paint.set_color(Color::from_rgba8(200, 200, 0, 255));
-    for (i, dim) in data.dimensions.iter().enumerate() {
-        let selected = state.selected_dimensions.borrow().contains(&i);
-        if selected {
-            paint.set_color(Color::from_rgba8(255, 255, 0, 255));
-        } else {
-            paint.set_color(Color::from_rgba8(200, 200, 0, 255));
-        }
-        let mut pb = PathBuilder::new();
-        pb.move_to(tx(dim.start.x as f32), ty(dim.start.y as f32));
-        pb.line_to(tx(dim.end.x as f32), ty(dim.end.y as f32));
-        if let Some(path) = pb.finish() {
-            let stroke = Stroke {
-                width: 1.0,
-                ..Stroke::default()
-            };
-            pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-        }
-        let line = Line::new(dim.start, dim.end);
-        let mid = line.midpoint();
-        let text = if let Some(t) = &dim.text {
-            t.clone()
-        } else {
-            format!("{:.2}", line.length())
-        };
-        draw_text(
-            &mut pixmap,
-            &text,
-            &FONT,
-            tx(mid.x as f32),
-            ty(mid.y as f32 - 10.0),
-            Color::from_rgba8(255, 255, 255, 255),
-            12.0,
-        );
-    }
-
-    paint.set_color(Color::from_rgba8(128, 128, 128, 255));
-    for tin in data.surfaces {
-        for tri in &tin.triangles {
-            let a = tin.vertices[tri[0]];
-            let b = tin.vertices[tri[1]];
-            let c = tin.vertices[tri[2]];
-            let mut pb = PathBuilder::new();
-            pb.move_to(tx(a.x as f32), ty(a.y as f32));
-            pb.line_to(tx(b.x as f32), ty(b.y as f32));
-            pb.line_to(tx(c.x as f32), ty(c.y as f32));
-            pb.close();
-            if let Some(path) = pb.finish() {
-                let stroke = Stroke {
-                    width: 1.0,
-                    ..Stroke::default()
-                };
-                pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-            }
-        }
-    }
-
-    paint.set_color(Color::from_rgba8(
-        styles.alignment_style.color[0],
-        styles.alignment_style.color[1],
-        styles.alignment_style.color[2],
-        255,
-    ));
-    for al in data.alignments {
-        for elem in &al.horizontal.elements {
-            match elem {
-                survey_cad::alignment::HorizontalElement::Tangent { start, end } => {
-                    let mut pb = PathBuilder::new();
-                    pb.move_to(tx(start.x as f32), ty(start.y as f32));
-                    pb.line_to(tx(end.x as f32), ty(end.y as f32));
-                    if let Some(path) = pb.finish() {
-                        let mut stroke = Stroke {
-                            width: styles.alignment_style.weight.0,
-                            ..Stroke::default()
-                        };
-                        use tiny_skia::StrokeDash;
-                        match styles.alignment_style.line_type {
-                            LineType::Dashed => {
-                                stroke.dash = StrokeDash::new(vec![10.0, 10.0], 0.0);
-                            }
-                            LineType::Dotted => {
-                                stroke.dash = StrokeDash::new(vec![2.0, 6.0], 0.0);
-                            }
-                            _ => {}
-                        }
-                        pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-                    }
-                }
-                survey_cad::alignment::HorizontalElement::Curve { arc } => {
-                    let steps = 32;
-                    let mut pb = PathBuilder::new();
-                    for i in 0..=steps {
-                        let t = arc.start_angle
-                            + (arc.end_angle - arc.start_angle) * (i as f64 / steps as f64);
-                        let x = arc.center.x + arc.radius * t.cos();
-                        let y = arc.center.y + arc.radius * t.sin();
-                        let px = tx(x as f32);
-                        let py = ty(y as f32);
-                        if i == 0 {
-                            pb.move_to(px, py);
-                        } else {
-                            pb.line_to(px, py);
-                        }
-                    }
-                    if let Some(path) = pb.finish() {
-                        let mut stroke = Stroke {
-                            width: styles.alignment_style.weight.0,
-                            ..Stroke::default()
-                        };
-                        use tiny_skia::StrokeDash;
-                        match styles.alignment_style.line_type {
-                            LineType::Dashed => {
-                                stroke.dash = StrokeDash::new(vec![10.0, 10.0], 0.0);
-                            }
-                            LineType::Dotted => {
-                                stroke.dash = StrokeDash::new(vec![2.0, 6.0], 0.0);
-                            }
-                            _ => {}
-                        }
-                        pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-                    }
-                }
-                survey_cad::alignment::HorizontalElement::Spiral { spiral } => {
-                    let mut pb = PathBuilder::new();
-                    let sp = spiral.start_point();
-                    let ep = spiral.end_point();
-                    pb.move_to(tx(sp.x as f32), ty(sp.y as f32));
-                    pb.line_to(tx(ep.x as f32), ty(ep.y as f32));
-                    if let Some(path) = pb.finish() {
-                        let mut stroke = Stroke {
-                            width: styles.alignment_style.weight.0,
-                            ..Stroke::default()
-                        };
-                        use tiny_skia::StrokeDash;
-                        match styles.alignment_style.line_type {
-                            LineType::Dashed => {
-                                stroke.dash = StrokeDash::new(vec![10.0, 10.0], 0.0);
-                            }
-                            LineType::Dotted => {
-                                stroke.dash = StrokeDash::new(vec![2.0, 6.0], 0.0);
-                            }
-                            _ => {}
-                        }
-                        pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-                    }
-                }
-            }
-        }
-    }
-
-    for (idx, p) in data.points.iter().enumerate() {
-        let style_idx = styles.style_indices.borrow().get(idx).copied().unwrap_or(0);
-        let style = styles
-            .point_styles
-            .get(style_idx)
-            .copied()
-            .unwrap_or(PointStyle::new(PointSymbol::Circle, [0, 255, 0], 3.0));
-        let selected = state.selected.borrow().contains(&idx);
-        let color = if selected { [255, 255, 0] } else { style.color };
-        paint.set_color(Color::from_rgba8(color[0], color[1], color[2], 255));
-        match style.symbol {
-            PointSymbol::Circle => {
-                if let Some(c) =
-                    PathBuilder::from_circle(tx(p.x as f32), ty(p.y as f32), style.size)
-                {
-                    pixmap.fill_path(
-                        &c,
-                        &paint,
-                        tiny_skia::FillRule::Winding,
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            PointSymbol::Square => {
-                let half = style.size;
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(p.x as f32 - half), ty(p.y as f32 - half));
-                pb.line_to(tx(p.x as f32 + half), ty(p.y as f32 - half));
-                pb.line_to(tx(p.x as f32 + half), ty(p.y as f32 + half));
-                pb.line_to(tx(p.x as f32 - half), ty(p.y as f32 + half));
-                pb.close();
-                if let Some(path) = pb.finish() {
-                    pixmap.fill_path(
-                        &path,
-                        &paint,
-                        tiny_skia::FillRule::Winding,
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            PointSymbol::Cross => {
-                let half = style.size;
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(p.x as f32 - half), ty(p.y as f32 - half));
-                pb.line_to(tx(p.x as f32 + half), ty(p.y as f32 + half));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(p.x as f32 - half), ty(p.y as f32 + half));
-                pb.line_to(tx(p.x as f32 + half), ty(p.y as f32 - half));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-        }
-        if styles.show_point_numbers {
-            draw_text(
-                &mut pixmap,
-                &(idx + 1).to_string(),
-                &FONT,
-                tx(p.x as f32 + styles.point_label_style.offset[0]),
-                ty(p.y as f32 + styles.point_label_style.offset[1]),
-                Color::from_rgba8(
-                    styles.point_label_style.color[0],
-                    styles.point_label_style.color[1],
-                    styles.point_label_style.color[2],
-                    255,
-                ),
-                styles.point_label_style.text_style.height as f32,
-            );
-        }
-    }
-
-    if let Some(cf) = state.cursor_feedback.borrow().as_ref() {
-        let wp = screen_to_workspace(
-            cf.pos.0,
-            cf.pos.1,
-            state.offset,
-            state.zoom,
-            width as f32,
-            height as f32,
-        );
-        paint.set_color(Color::from_rgba8(255, 255, 0, 255));
-        match drawing {
-            DrawingMode::Line { start: Some(s) } => {
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(s.x as f32), ty(s.y as f32));
-                pb.line_to(tx(wp.x as f32), ty(wp.y as f32));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            DrawingMode::Dimension { start: Some(s) } => {
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(s.x as f32), ty(s.y as f32));
-                pb.line_to(tx(wp.x as f32), ty(wp.y as f32));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            DrawingMode::Polygon { vertices } if !vertices.is_empty() => {
-                let mut pb = PathBuilder::new();
-                let Some(first) = vertices.first() else { return Ok(Image::default()); };
-                pb.move_to(tx(first.x as f32), ty(first.y as f32));
-                for p in &vertices[1..] {
-                    pb.line_to(tx(p.x as f32), ty(p.y as f32));
-                }
-                pb.line_to(tx(wp.x as f32), ty(wp.y as f32));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            DrawingMode::ArcCenter {
-                center: Some(c),
-                radius: Some(r),
-                start_angle: Some(sa),
-            } => {
-                let ea = (wp.y - c.y).atan2(wp.x - c.x);
-                let mut pb = PathBuilder::new();
-                for i in 0..=32 {
-                    let t = sa + (ea - sa) * (i as f64 / 32.0);
-                    let x = c.x + r * t.cos();
-                    let y = c.y + r * t.sin();
-                    if i == 0 {
-                        pb.move_to(tx(x as f32), ty(y as f32));
-                    } else {
-                        pb.line_to(tx(x as f32), ty(y as f32));
-                    }
-                }
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            DrawingMode::ArcThreePoint {
-                p1: Some(p1),
-                p2: None,
-            } => {
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(p1.x as f32), ty(p1.y as f32));
-                pb.line_to(tx(wp.x as f32), ty(wp.y as f32));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            DrawingMode::ArcThreePoint {
-                p1: Some(p1),
-                p2: Some(p2),
-            } => {
-                if let Some(arc) = arc_from_three_points(*p1, *p2, wp) {
-                    let mut pb = PathBuilder::new();
-                    for i in 0..=32 {
-                        let t =
-                            arc.start_angle + (arc.end_angle - arc.start_angle) * (i as f64 / 32.0);
-                        let x = arc.center.x + arc.radius * t.cos();
-                        let y = arc.center.y + arc.radius * t.sin();
-                        if i == 0 {
-                            pb.move_to(tx(x as f32), ty(y as f32));
-                        } else {
-                            pb.line_to(tx(x as f32), ty(y as f32));
-                        }
-                    }
-                    if let Some(path) = pb.finish() {
-                        pixmap.stroke_path(
-                            &path,
-                            &paint,
-                            &Stroke {
-                                width: 1.0,
-                                ..Stroke::default()
-                            },
-                            Transform::identity(),
-                            None,
-                        );
-                    }
-                }
-            }
-            DrawingMode::ArcStartEndRadius {
-                start: Some(s),
-                end: None,
-                ..
-            } => {
-                let mut pb = PathBuilder::new();
-                pb.move_to(tx(s.x as f32), ty(s.y as f32));
-                pb.line_to(tx(wp.x as f32), ty(wp.y as f32));
-                if let Some(path) = pb.finish() {
-                    pixmap.stroke_path(
-                        &path,
-                        &paint,
-                        &Stroke {
-                            width: 1.0,
-                            ..Stroke::default()
-                        },
-                        Transform::identity(),
-                        None,
-                    );
-                }
-            }
-            DrawingMode::ArcStartEndRadius {
-                start: Some(s),
-                end: Some(e),
-                radius: None,
-            } => {
-                let r = ((wp.x - s.x).powi(2) + (wp.y - s.y).powi(2)).sqrt();
-                if let Some(arc) = arc_from_start_end_radius(*s, *e, r, wp) {
-                    let mut pb = PathBuilder::new();
-                    for i in 0..=32 {
-                        let t =
-                            arc.start_angle + (arc.end_angle - arc.start_angle) * (i as f64 / 32.0);
-                        let x = arc.center.x + arc.radius * t.cos();
-                        let y = arc.center.y + arc.radius * t.sin();
-                        if i == 0 {
-                            pb.move_to(tx(x as f32), ty(y as f32));
-                        } else {
-                            pb.line_to(tx(x as f32), ty(y as f32));
-                        }
-                    }
-                    if let Some(path) = pb.finish() {
-                        pixmap.stroke_path(
-                            &path,
-                            &paint,
-                            &Stroke {
-                                width: 1.0,
-                                ..Stroke::default()
-                            },
-                            Transform::identity(),
-                            None,
-                        );
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    if let Some(sp) = state.snap_target.borrow().as_ref() {
-        paint.set_color(Color::from_rgba8(255, 0, 0, 255));
-        let r = 3.0;
-        let (sx, sy) = (tx(sp.x as f32), ty(sp.y as f32));
-        let mut pb = PathBuilder::new();
-        pb.move_to(sx - r, sy);
-        pb.line_to(sx + r, sy);
-        pb.move_to(sx, sy - r);
-        pb.line_to(sx, sy + r);
-        if let Some(path) = pb.finish() {
-            pixmap.stroke_path(
-                &path,
-                &paint,
-                &Stroke {
-                    width: 1.0,
-                    ..Stroke::default()
-                },
-                Transform::identity(),
-                None,
-            );
-        }
-    }
-
-    if state.drag.borrow().active {
-        let ds = state.drag.borrow();
-        let x1 = ds.start.0.min(ds.end.0);
-        let y1 = ds.start.1.min(ds.end.1);
-        let x2 = ds.start.0.max(ds.end.0);
-        let y2 = ds.start.1.max(ds.end.1);
-        paint.set_color(Color::from_rgba8(255, 255, 255, 128));
-        let rect_stroke = Stroke {
-            width: 1.0,
-            ..Stroke::default()
-        };
-        let mut pb = PathBuilder::new();
-        pb.move_to(x1, y1);
-        pb.line_to(x2, y1);
-        pb.line_to(x2, y2);
-        pb.line_to(x1, y2);
-        pb.close();
-        if let Some(path) = pb.finish() {
-            pixmap.stroke_path(&path, &paint, &rect_stroke, Transform::identity(), None);
-        }
-    }
-
-    if let Some(cf) = state.cursor_feedback.borrow().as_ref() {
-        let t = (cf.frame % 30) as f32 / 30.0;
-        paint.set_color(Color::from_rgba8(
-            (255.0 * t) as u8,
-            (255.0 * (1.0 - t)) as u8,
-            0,
-            255,
-        ));
-        let mut pb = PathBuilder::new();
-        pb.move_to(cf.pos.0 - 5.0, cf.pos.1);
-        pb.line_to(cf.pos.0 + 5.0, cf.pos.1);
-        pb.move_to(cf.pos.0, cf.pos.1 - 5.0);
-        pb.line_to(cf.pos.0, cf.pos.1 + 5.0);
-        if let Some(path) = pb.finish() {
-            pixmap.stroke_path(
-                &path,
-                &paint,
-                &Stroke {
-                    width: 1.0,
-                    ..Stroke::default()
-                },
-                Transform::identity(),
-                None,
-            );
-        }
-    }
-    let buffer = slint::SharedPixelBuffer::<slint::Rgba8Pixel>::clone_from_slice(
-        pixmap.data(),
-        width,
-        height,
-    );
-    Ok(Image::from_rgba8_premultiplied(buffer))
-}
-
-fn render_cross_section(
-    section: &corridor::CrossSection,
-    width: u32,
-    height: u32,
-) -> Result<Image, Box<dyn Error>> {
-    if width == 0 || height == 0 {
-        return Ok(Image::default());
-    }
-    let mut pixmap = Pixmap::new(width, height).ok_or("failed to create pixmap")?;
-    pixmap.fill(Color::from_rgba8(32, 32, 32, 255));
-    let mut paint = Paint::default();
-    paint.set_color(Color::from_rgba8(0, 255, 0, 255));
-    paint.anti_alias = true;
-
-    if section.points.len() >= 2 {
-        let Some(first) = section.points.first() else { return Ok(Image::default()); };
-        let Some(last) = section.points.last() else { return Ok(Image::default()); };
-        let dx = last.x - first.x;
-        let dy = last.y - first.y;
-        let len = (dx * dx + dy * dy).sqrt();
-        let dir = if len.abs() < f64::EPSILON {
-            (1.0, 0.0)
-        } else {
-            (dx / len, dy / len)
-        };
-        let center = section.points[section.points.len() / 2];
-        let mut pts = Vec::new();
-        let mut min_x = f32::MAX;
-        let mut max_x = f32::MIN;
-        let mut min_y = f32::MAX;
-        let mut max_y = f32::MIN;
-        for p in &section.points {
-            let off = ((p.x - center.x) * dir.0 + (p.y - center.y) * dir.1) as f32;
-            let elev = (p.z - center.z) as f32;
-            pts.push((off, elev));
-            min_x = min_x.min(off);
-            max_x = max_x.max(off);
-            min_y = min_y.min(elev);
-            max_y = max_y.max(elev);
-        }
-        if (max_x - min_x).abs() < f32::EPSILON {
-            max_x += 1.0;
-        }
-        if (max_y - min_y).abs() < f32::EPSILON {
-            max_y += 1.0;
-        }
-        let scale =
-            ((width as f32 * 0.8) / (max_x - min_x)).min((height as f32 * 0.8) / (max_y - min_y));
-        let ox = width as f32 / 2.0 - scale * (min_x + max_x) / 2.0;
-        let oy = height as f32 / 2.0 + scale * (min_y + max_y) / 2.0;
-        let mut pb = PathBuilder::new();
-        for (i, (x, y)) in pts.iter().enumerate() {
-            let px = ox + *x * scale;
-            let py = oy - *y * scale;
-            if i == 0 {
-                pb.move_to(px, py);
-            } else {
-                pb.line_to(px, py);
-            }
-        }
-        if let Some(path) = pb.finish() {
-            let stroke = Stroke {
-                width: 2.0,
-                ..Stroke::default()
-            };
-            pixmap.stroke_path(&path, &paint, &stroke, Transform::identity(), None);
-        }
-    }
-
-    let buffer = slint::SharedPixelBuffer::<slint::Rgba8Pixel>::clone_from_slice(
-        pixmap.data(),
-        width,
-        height,
-    );
-    Ok(Image::from_rgba8_premultiplied(buffer))
-}
-
-struct SectionParams {
-    dir: (f64, f64),
-    center: ScPoint3,
-    scale: f32,
-    ox: f32,
-    oy: f32,
-}
-
-fn calc_section_params(
-    section: &corridor::CrossSection,
-    width: f32,
-    height: f32,
-) -> Option<SectionParams> {
-    if section.points.len() < 2 {
-        return None;
-    }
-    let first = section.points.first()?;
-    let last = section.points.last()?;
-    let dx = last.x - first.x;
-    let dy = last.y - first.y;
-    let len = (dx * dx + dy * dy).sqrt();
-    let dir = if len.abs() < f64::EPSILON {
-        (1.0, 0.0)
-    } else {
-        (dx / len, dy / len)
-    };
-    let center = section.points[section.points.len() / 2];
-    let mut min_x = f32::MAX;
-    let mut max_x = f32::MIN;
-    let mut min_y = f32::MAX;
-    let mut max_y = f32::MIN;
-    for p in &section.points {
-        let off = ((p.x - center.x) * dir.0 + (p.y - center.y) * dir.1) as f32;
-        let elev = (p.z - center.z) as f32;
-        min_x = min_x.min(off);
-        max_x = max_x.max(off);
-        min_y = min_y.min(elev);
-        max_y = max_y.max(elev);
-    }
-    if (max_x - min_x).abs() < f32::EPSILON {
-        max_x += 1.0;
-    }
-    if (max_y - min_y).abs() < f32::EPSILON {
-        max_y += 1.0;
-    }
-    let scale = ((width * 0.8) / (max_x - min_x)).min((height * 0.8) / (max_y - min_y));
-    let ox = width / 2.0 - scale * (min_x + max_x) / 2.0;
-    let oy = height / 2.0 + scale * (min_y + max_y) / 2.0;
-    Some(SectionParams {
-        dir,
-        center,
-        scale,
-        ox,
-        oy,
-    })
-}
-
-fn screen_to_world(
-    section: &corridor::CrossSection,
-    x: f32,
-    y: f32,
-    width: f32,
-    height: f32,
-) -> Option<ScPoint3> {
-    let params = calc_section_params(section, width, height)?;
-    let off = (x - params.ox) / params.scale;
-    let elev = (params.oy - y) / params.scale;
-    Some(ScPoint3::new(
-        params.center.x + off as f64 * params.dir.0,
-        params.center.y + off as f64 * params.dir.1,
-        params.center.z + elev as f64,
-    ))
-}
-
-fn nearest_point(
-    section: &corridor::CrossSection,
-    x: f32,
-    y: f32,
-    width: f32,
-    height: f32,
-) -> Option<usize> {
-    let params = calc_section_params(section, width, height)?;
-    let mut best = None;
-    let mut best_dist = f32::MAX;
-    for (i, p) in section.points.iter().enumerate() {
-        let off = ((p.x - params.center.x) * params.dir.0 + (p.y - params.center.y) * params.dir.1)
-            as f32;
-        let elev = (p.z - params.center.z) as f32;
-        let sx = params.ox + off * params.scale;
-        let sy = params.oy - elev * params.scale;
-        let dx = sx - x;
-        let dy = sy - y;
-        let dist = dx * dx + dy * dy;
-        if dist < best_dist {
-            best_dist = dist;
-            best = Some(i);
-        }
-    }
-    if best_dist.sqrt() <= 10.0 {
-        best
-    } else {
-        None
-    }
-}
-
-fn grade_at(profile: &VerticalAlignment, station: f64) -> Option<f64> {
-    for elem in &profile.elements {
-        match *elem {
-            VerticalElement::Grade {
-                start_station,
-                end_station,
-                start_elev,
-                end_elev,
-            } => {
-                if station >= start_station && station <= end_station {
-                    if (end_station - start_station).abs() < f64::EPSILON {
-                        return Some(0.0);
-                    }
-                    return Some((end_elev - start_elev) / (end_station - start_station));
-                }
-            }
-            VerticalElement::Parabola {
-                start_station,
-                end_station,
-                start_grade,
-                end_grade,
-                ..
-            } => {
-                if station >= start_station && station <= end_station {
-                    let t = (station - start_station) / (end_station - start_station);
-                    return Some(start_grade + (end_grade - start_grade) * t);
-                }
-            }
-        }
-    }
-    None
-}
-
-fn read_line_csv(path: &str, dst_epsg: u32) -> std::io::Result<(Point, Point)> {
-    let pts = survey_cad::io::read_points_csv(path, Some(4326), Some(dst_epsg))?;
-    if pts.len() != 2 {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            "expected exactly two points",
-        ));
-    }
-    Ok((pts[0], pts[1]))
-}
-
-fn read_points_list(path: &str, dst_epsg: u32) -> std::io::Result<Vec<Point>> {
-    survey_cad::io::read_points_csv(path, Some(4326), Some(dst_epsg))
-}
-
-fn read_arc_csv(path: &str) -> std::io::Result<Arc> {
-    let lines = survey_cad::io::read_lines(path)?;
-    if lines.is_empty() {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            "empty file",
-        ));
-    }
-    let parts: Vec<&str> = lines[0].split(',').collect();
-    if parts.len() != 5 {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            "expected cx,cy,radius,start,end",
-        ));
-    }
-    let cx: f64 = parts[0]
-        .trim()
-        .parse()
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-    let cy: f64 = parts[1]
-        .trim()
-        .parse()
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-    let r: f64 = parts[2]
-        .trim()
-        .parse()
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-    let sa: f64 = parts[3]
-        .trim()
-        .parse()
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-    let ea: f64 = parts[4]
-        .trim()
-        .parse()
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-    Ok(Arc::new(Point::new(cx, cy), r, sa, ea))
-}
-
-fn refresh_workspace(
-    app: &MainWindow,
-    render_image: &dyn Fn() -> Result<Image, Box<dyn Error>>,
-    backend_render: &Rc<RefCell<TruckBackend>>,
-) {
-    if app.get_workspace_mode() == 0 {
-        match render_image() {
-            Ok(img) => app.set_workspace_image(img),
-            Err(e) => {
-                eprintln!("Failed to render workspace: {e}");
-                app.set_workspace_image(Image::default());
-            }
-        }
-    } else {
-        let image = backend_render.borrow_mut().render();
-        app.set_workspace_texture(image);
-    }
-    app.window().request_redraw();
-}
-
-pub fn set_workspace_image_result(
-    app: &MainWindow,
-    render_image: &dyn Fn() -> Result<Image, Box<dyn Error>>,
-) {
-    match render_image() {
-        Ok(img) => app.set_workspace_image(img),
-        Err(e) => {
-            eprintln!("Failed to render workspace: {e}");
-            app.set_workspace_image(Image::default());
-        }
-    }
-}
-
-fn show_context_menu(
-    app: &MainWindow,
-    state: &Rc<RefCell<Option<slint::Weak<ContextMenu>>>>,
-    x: f32,
-    y: f32,
-) {
-    if let Some(m) = state.borrow_mut().take().and_then(|w| w.upgrade()) {
-        let _ = m.hide();
-    }
-    let menu = ContextMenu::new().unwrap();
-    menu.set_pos_x(x);
-    menu.set_pos_y(y);
-    {
-        let weak = app.as_weak();
-        menu.on_mov(move || {
-            if let Some(a) = weak.upgrade() {
-                a.invoke_move_entity();
-            }
-        });
-    }
-
-    {
-        let weak = app.as_weak();
-        menu.on_rot(move || {
-            if let Some(a) = weak.upgrade() {
-                a.invoke_rotate_entity();
-            }
-        });
-    }
-    {
-        let weak = app.as_weak();
-        menu.on_properties(move || {
-            if let Some(a) = weak.upgrade() {
-                a.invoke_inspector();
-            }
-        });
-    }
-    {
-        let weak = app.as_weak();
-        menu.on_delete(move || {
-            if let Some(a) = weak.upgrade() {
-                a.invoke_delete_selected();
-            }
-        });
-    }
-    menu.show().unwrap();
-    *state.borrow_mut() = Some(menu.as_weak());
-}
-
-fn has_selection(
-    pts: &Rc<RefCell<Vec<usize>>>,
-    lines: &Rc<RefCell<Vec<(Point, Point)>>>,
-    polys: &Rc<RefCell<Vec<usize>>>,
-    plines: &Rc<RefCell<Vec<usize>>>,
-    arcs: &Rc<RefCell<Vec<usize>>>,
-    dims: &Rc<RefCell<Vec<usize>>>,
-) -> bool {
-    !pts.borrow().is_empty()
-        || !lines.borrow().is_empty()
-        || !polys.borrow().is_empty()
-        || !plines.borrow().is_empty()
-        || !arcs.borrow().is_empty()
-        || !dims.borrow().is_empty()
-}
-
-#[allow(clippy::too_many_arguments)]
-fn show_inspector_for_point(
-    idx: usize,
-    app: &MainWindow,
-    layer_names: &Rc<RefCell<Vec<String>>>,
-    style_names: &[SharedString],
-    layers: &Rc<RefCell<Vec<usize>>>,
-    styles: &Rc<RefCell<Vec<usize>>>,
-    metadata: &Rc<RefCell<Vec<String>>>,
-    elevation: &Rc<RefCell<Vec<String>>>,
-    measurement: &Rc<RefCell<Vec<String>>>,
-    data_sets: &Rc<RefCell<Vec<usize>>>,
-    data_set_names: &Rc<RefCell<Vec<String>>>,
-    inspector: &Rc<RefCell<Option<slint::Weak<EntityInspector>>>>,
-    render_image: Rc<dyn Fn() -> Result<Image, Box<dyn Error>>>,
-    backend: &Rc<RefCell<TruckBackend>>,
-) {
-    while layers.borrow().len() <= idx {
-        layers.borrow_mut().push(0);
-    }
-    while styles.borrow().len() <= idx {
-        styles.borrow_mut().push(0);
-    }
-    while metadata.borrow().len() <= idx {
-        metadata.borrow_mut().push(String::new());
-    }
-    while elevation.borrow().len() <= idx {
-        elevation.borrow_mut().push(String::new());
-    }
-    while measurement.borrow().len() <= idx {
-        measurement.borrow_mut().push(String::new());
-    }
-    while data_sets.borrow().len() <= idx {
-        data_sets.borrow_mut().push(0);
-    }
-
-    let layer_model = Rc::new(VecModel::from(
-        layer_names
-            .borrow()
-            .iter()
-            .cloned()
-            .map(SharedString::from)
-            .collect::<Vec<_>>(),
-    ));
-    let style_model = Rc::new(VecModel::from(style_names.to_vec()));
-    let data_set_model = Rc::new(VecModel::from(
-        data_set_names
-            .borrow()
-            .iter()
-            .cloned()
-            .map(SharedString::from)
-            .collect::<Vec<_>>(),
-    ));
-
-    let dlg = if let Some(w) = inspector.borrow().as_ref().and_then(|w| w.upgrade()) {
-        w
-    } else {
-        let d = EntityInspector::new().unwrap();
-        *inspector.borrow_mut() = Some(d.as_weak());
-        d
-    };
-
-    dlg.set_layers_model(layer_model.into());
-    dlg.set_styles_model(style_model.into());
-    dlg.set_data_set_model(data_set_model.into());
-    dlg.set_entity_type(SharedString::from("Point"));
-    dlg.set_layer_index(layers.borrow()[idx] as i32);
-    dlg.set_style_index(styles.borrow()[idx] as i32);
-    dlg.set_metadata(SharedString::from(metadata.borrow()[idx].clone()));
-    dlg.set_elevation(SharedString::from(elevation.borrow()[idx].clone()));
-    dlg.set_measurement(SharedString::from(measurement.borrow()[idx].clone()));
-    dlg.set_data_set_index(data_sets.borrow()[idx] as i32);
-
-    {
-        let layers = layers.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_layer_changed(move |val| {
-            if let Some(l) = layers.borrow_mut().get_mut(idx) {
-                *l = val as usize;
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let styles_ref = styles.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_style_changed(move |val| {
-            if let Some(s) = styles_ref.borrow_mut().get_mut(idx) {
-                *s = val as usize;
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let meta_ref = metadata.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_metadata_changed(move |text| {
-            if let Some(m) = meta_ref.borrow_mut().get_mut(idx) {
-                *m = text.to_string();
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let elev_ref = elevation.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_elevation_changed(move |text| {
-            if let Some(e) = elev_ref.borrow_mut().get_mut(idx) {
-                *e = text.to_string();
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let meas_ref = measurement.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_measurement_changed(move |text| {
-            if let Some(m) = meas_ref.borrow_mut().get_mut(idx) {
-                *m = text.to_string();
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let ds_ref = data_sets.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_data_set_changed(move |val| {
-            if let Some(d) = ds_ref.borrow_mut().get_mut(idx) {
-                *d = val as usize;
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    dlg.show().unwrap();
-}
-
-#[allow(clippy::too_many_arguments)]
-fn show_inspector_for_polygon(
-    idx: usize,
-    app: &MainWindow,
-    layer_names: &Rc<RefCell<Vec<String>>>,
-    hatch_names: &[SharedString],
-    layers: &Rc<RefCell<Vec<usize>>>,
-    hatches: &Rc<RefCell<Vec<usize>>>,
-    measurement: &Rc<RefCell<Vec<String>>>,
-    data_sets: &Rc<RefCell<Vec<usize>>>,
-    data_set_names: &Rc<RefCell<Vec<String>>>,
-    inspector: &Rc<RefCell<Option<slint::Weak<EntityInspector>>>>,
-    render_image: Rc<dyn Fn() -> Result<Image, Box<dyn Error>>>,
-    backend: &Rc<RefCell<TruckBackend>>,
-) {
-    while layers.borrow().len() <= idx {
-        layers.borrow_mut().push(0);
-    }
-    while hatches.borrow().len() <= idx {
-        hatches.borrow_mut().push(0);
-    }
-    while measurement.borrow().len() <= idx {
-        measurement.borrow_mut().push(String::new());
-    }
-    while data_sets.borrow().len() <= idx {
-        data_sets.borrow_mut().push(0);
-    }
-
-    let layer_model = Rc::new(VecModel::from(
-        layer_names
-            .borrow()
-            .iter()
-            .cloned()
-            .map(SharedString::from)
-            .collect::<Vec<_>>(),
-    ));
-    let hatch_model = Rc::new(VecModel::from(hatch_names.to_vec()));
-    let data_set_model = Rc::new(VecModel::from(
-        data_set_names
-            .borrow()
-            .iter()
-            .cloned()
-            .map(SharedString::from)
-            .collect::<Vec<_>>(),
-    ));
-
-    let dlg = if let Some(w) = inspector.borrow().as_ref().and_then(|w| w.upgrade()) {
-        w
-    } else {
-        let d = EntityInspector::new().unwrap();
-        *inspector.borrow_mut() = Some(d.as_weak());
-        d
-    };
-
-    dlg.set_layers_model(layer_model.into());
-    dlg.set_styles_model(Rc::new(VecModel::from(Vec::<SharedString>::new())).into());
-    dlg.set_hatch_model(hatch_model.into());
-    dlg.set_data_set_model(data_set_model.into());
-    dlg.set_entity_type(SharedString::from("Polygon"));
-    dlg.set_layer_index(layers.borrow()[idx] as i32);
-    dlg.set_hatch_index(hatches.borrow()[idx] as i32);
-    dlg.set_metadata(SharedString::from(""));
-    dlg.set_measurement(SharedString::from(measurement.borrow()[idx].clone()));
-    dlg.set_data_set_index(data_sets.borrow()[idx] as i32);
-
-    {
-        let layers = layers.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_layer_changed(move |val| {
-            if let Some(l) = layers.borrow_mut().get_mut(idx) {
-                *l = val as usize;
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let h_ref = hatches.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_hatch_changed(move |val| {
-            if let Some(h) = h_ref.borrow_mut().get_mut(idx) {
-                *h = val as usize;
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let meas_ref = measurement.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_measurement_changed(move |text| {
-            if let Some(m) = meas_ref.borrow_mut().get_mut(idx) {
-                *m = text.to_string();
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    {
-        let ds_ref = data_sets.clone();
-        let app_weak = app.as_weak();
-        let backend = backend.clone();
-        let render_image = render_image.clone();
-        dlg.on_data_set_changed(move |val| {
-            if let Some(d) = ds_ref.borrow_mut().get_mut(idx) {
-                *d = val as usize;
-            }
-            if let Some(a) = app_weak.upgrade() {
-                refresh_workspace(&a, &*render_image, &backend);
-            }
-        });
-    }
-
-    dlg.show().unwrap();
-}
 
 fn main() -> Result<(), slint::PlatformError> {
     let mut cmd_font: Option<String> = None;
